https://github.com/krishnasaye23/CollegeAPI/issues/4#issue-910215271

+++ src/main/java/com/example/schoolapi/Service/DepartmentsServiceImpl.java
@@ -1,37 +1,59 @@
 package com.example.schoolapi.Service;
 
 import com.example.schoolapi.Entity.DepartmentEntity;
+import com.example.schoolapi.Model.Departments;
 import com.example.schoolapi.Repository.DepartmentRepo;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
+
 @Service
 public class DepartmentsServiceImpl implements DepartmentsService {
     @Autowired
     private DepartmentRepo departmentRepo;
+    @Autowired
+    private ModelMapper modelMapper;
     @Override
-    public List<DepartmentEntity> getDepartments() {
-        return departmentRepo.findAll();
+    public List<Departments> getDepartments() {
+
+        return departmentRepo.findAll().stream().map(departmentEntity -> modelMapper.map(departmentEntity,Departments.class))
+                .collect(Collectors.toList());
     }
     @Override
-    public DepartmentEntity getDepartmentbyid(int dept_id) {
-        DepartmentEntity opt=departmentRepo.findById(dept_id);
-        return opt;
+    public Departments getDepartmentbyid(int dept_id) {
+        DepartmentEntity departmentEntity= departmentRepo.findById(dept_id);
+        Departments dep=modelMapper.map(departmentEntity,Departments.class);
+        return dep;
+
     }
     @Override
-    public DepartmentEntity addDepartments(DepartmentEntity department){
-        departmentRepo.save(department);
-        return department;
+    public Departments addDepartments(Departments department){
+        DepartmentEntity dep=modelMapper.map(department,DepartmentEntity.class);
+        departmentRepo.save(dep);
+        Departments deps=modelMapper.map(dep,Departments.class);
+        return deps;
+        /*DepartmentEntity departmentEntity=new DepartmentEntity();
+        departmentEntity.setDept_id(department.getDept_id());
+        departmentRepo.save(departmentEntity);
+        return department;*/
     }
     @Override
-    public DepartmentEntity updateDepartments(DepartmentEntity department){
-        departmentRepo.save(department);
-        return department;
+    public Departments updateDepartments(Departments department){
+        DepartmentEntity dep=modelMapper.map(department,DepartmentEntity.class);
+        departmentRepo.save(dep);
+        Departments deps=modelMapper.map(dep,Departments.class);
+        return deps;
     }
     @Override
-    public DepartmentEntity delrecord(int dept_id) {
-        DepartmentEntity g=departmentRepo.deleteById(dept_id);
-        return g;
+    public String delrecord(int dept_id) {
+        if(departmentRepo.findById(dept_id)!=null) {
+            departmentRepo.deleteById(dept_id);
+            return "Deleted successfully";
+        }
+        else
+            return "Record not found with the id";
     }
 }
